# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Document:
    properties:
      ai_analysis_json:
        type: object
      category:
        type: string
      created_at:
        type: string
      id:
        type: string
      last_updated_at:
        type: string
      source_uri:
        type: string
      status:
        type: string
      tenant_id:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
...
  handler.ConfirmUploadRequest:
    properties:
      category:
        type: string
      object_key:
        type: string
      title:
        type: string
    required:
    - category
    - object_key
    - title
    type: object
...
  handler.DocumentListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Document'
        type: array
      meta:
        $ref: '#/definitions/handler.Meta'
    type: object
  handler.ErrorResponse:
    properties:
      details:
        items:
          type: string
        type: array
      error:
        type: string
    type: object
...
  handler.Meta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:7777
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Elysian Backend API provides user authentication, management, and health
    check endpoints. Built with Go and Gin framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Elysian Backend API
  version: 1.0.0
paths:
...
  /api/v1/documents:
    get:
      description: List documents for tenant
      parameters:
      - description: Tenant ID (Strict UUID)
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Limit (max 100, default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DocumentListResponse'
        "400":
          description: Bad Request (Invalid X-Tenant-ID UUID format)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error (Sanitized, anti Information Disclosure)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List documents
      tags:
      - documents
  /api/v1/documents/confirm:
    post:
      consumes:
      - application/json
      description: Confirm S3 upload and trigger Asynq vectorization worker
      parameters:
      - description: Tenant ID (Strict UUID)
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Confirm Upload Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ConfirmUploadRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            properties:
              document_id:
                type: string
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error (Sanitized)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Confirm upload
      tags:
      - documents
  /api/v1/documents/presign:
    get:
      description: Get a 15-minute presigned S3 URL for direct file upload
      parameters:
      - description: Tenant ID (Strict UUID)
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Original filename
        in: query
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              expires_in:
                type: string
              object_key:
                type: string
              presigned_url:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error (Sanitized)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get presigned URL
      tags:
      - documents
  /api/v1/documents/{id}:
    delete:
      description: Deletes a document record from the database
      parameters:
      - description: Tenant ID (Strict UUID)
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error (Sanitized)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete document
      tags:
      - documents
  /api/v1/documents/{id}/approve:
    post:
      description: Transition status from pending_qa to processing and trigger chunk embedding
      parameters:
      - description: Tenant ID (Strict UUID)
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            properties:
              document_id:
                type: string
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error (Sanitized)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve document
      tags:
      - documents
  /api/v1/documents/{id}/raw:
    get:
      description: Returns the full raw text and its SHA-256 hash from MongoDB staging area
      parameters:
      - description: Tenant ID (Strict UUID)
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Document ID (UUID or Draft Name)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              hash:
                type: string
              id:
                type: string
              raw_text:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error (Sanitized)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get raw document text
      tags:
      - documents
  /api/v1/documents/{id}/text:
    patch:
      consumes:
      - application/json
      description: Updates the extracted plain text inside document's ai_analysis_json
      parameters:
      - description: Tenant ID (Strict UUID)
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Document ID (UUID or Draft Name)
        in: path
        name: id
        required: true
        type: string
      - description: Update text request body
        in: body
        name: request
        required: true
        schema:
          properties:
            extracted_text:
              type: string
            status:
              type: string
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error (Sanitized)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update document text
      tags:
      - documents
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"